+++ CHANGELOG.md
@@ -4,5 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.1 - 2021-03-04
+### Added
+- macOS build.
+### Changed
+- Installer code to be compitable with macOS.
+- Renamed PlansConverter to Plans Converter.
+### Fixed
+- Renamed delete output file to delete input file.
+- Extra "writing into" log message with FMS converter.
+
 ## 1.0 - 2021-02-26
 Initial release.